refactor: move css-sanitizer to src/sanitize/css
